refactor(swap): extract script/verify/derive helpers out of rfq.ts
rfq.ts should contain only transport and comms logic. Moves:
- VTXO script builders → vhtlcScript.ts
- Verification/assertion helpers → rfqVerify.ts
- Corridor derivation functions → rfqDerive.ts

No exported API changes.
